backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.6
django-cors-headers==4.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
